troubleshooting mean strange saving

+++ reshape_mrois_to_planes.py
@@ -191,7 +191,7 @@ if p['save']['mean']:
 
     print('volume min: {}'.format(np.min(volume)))
     print('volume max: {}'.format(np.max(volume)))
-    volume_mean = np.mean(volume, axis=0).astype(np.int16)
+    volume_mean = np.mean(volume, axis=0)  # .astype(np.int16)
     print('volume_mean dtype: {}'.format(volume_mean.dtype))
     pl, ph = np.percentile(volume_mean, [1, 99.9])
     volume_mean_rescale = img_as_ubyte(rescale_intensity(volume_mean, in_range=(pl, ph)))
